Add ushort local support for smooth 16-bit scrolling
- Pre-scan IL for conv.u2+stloc patterns to detect word-sized locals
- WriteStloc/WriteLdloc handle IsWord locals with 2-byte zero page storage
- HandleAddSub emits 16-bit INC (INC lo; BNE +3; INC hi) for word x++
- Conv_u1/u2/u4/i1/i2/i4 opcodes already handled as no-ops
- Change horizmask x from byte to ushort for smooth infinite scrolling
  (was wrapping at 256 pixels, now wraps at 512 across both nametables)

+++ samples/horizmask/Program.cs
@@ -23,7 +23,7 @@ void scroll_demo()
     byte bldg_height = (byte)((rand8() & 7) + 2);
     byte bldg_width = (byte)((rand8() & 3) * 4 + 4);
     byte bldg_char = (byte)(rand8() & 15);
-    byte x = 0;
+    ushort x = 0;
     byte col_counter = 32; // next column to draw (tile units)
     byte sub = 0; // sub-pixel counter (0-7)
     byte[] buf = new byte[PLAYROWS];
